% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_data_segments.R
\name{assemble_data_segments}
\alias{assemble_data_segments}
\title{Stitch together segments of time series measurements}
\usage{
assemble_data_segments(xlsx_file, segments = NULL)
}
\arguments{
\item{xlsx_file}{(character) path to Excel file}

\item{segments}{(numeric) number of segments to stitch together, starting
from the last sheet (default: number of sheets)}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} containing tidy data
}
\description{
Parses a Tecan Infinite 200 Pro plate reader Excel file that contains
segments of a time series spread over multiple sheets and automatically
stitches them together accounting for the time gap between them
}
\examples{
dat <- assemble_data_segments(
  system.file(
    "extdata",
    "tecan_time_series_segments.xlsx",
    package = "tecanr"
  )
)

}
