% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_types.R
\name{check_data_found}
\alias{check_data_found}
\title{Check whether plate reader data was found in Excel spreadsheet}
\usage{
check_data_found(x, y = "")
}
\arguments{
\item{x}{First value to check}

\item{y}{Second value to check}
}
\description{
Check whether plate reader data was found in Excel spreadsheet
}
\keyword{internal}
