% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_types.R
\name{check_file_exists}
\alias{check_file_exists}
\title{Check if file exists}
\usage{
check_file_exists(file)
}
\arguments{
\item{file}{Path to file}
}
\description{
Check if file exists
}
\keyword{internal}
