% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_time_single_reads.R
\name{single_time_single_reads}
\alias{single_time_single_reads}
\title{Read single measurement per well}
\usage{
single_time_single_reads(dat_raw)
}
\arguments{
\item{dat_raw}{(tibble) a excel sheet}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} containing tidy data
}
\description{
\code{\link[=single_time_single_reads]{single_time_single_reads()}} gets a Tecan plate reader measurement
from an Excel file when a single reading was taken of each well
}
\examples{
traw$single_time_single_reads |> single_time_single_reads()

}
