% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tecan_parse.R
\name{tecan_parse}
\alias{tecan_parse}
\title{Parses a Tecan Infinite 200 Pro plate reader excel file, determines the type
of measurement (single time point or time series, single or multiple reads
per well), and reads the data}
\usage{
tecan_parse(xlsx_file, xlsx_sheet = 1)
}
\arguments{
\item{xlsx_file}{(character) path to Excel file}

\item{xlsx_sheet}{(numeric) number of Excel sheet to read (default: 1)}
}
\value{
(tibble) long format data
}
\description{
Parses a Tecan Infinite 200 Pro plate reader excel file, determines the type
of measurement (single time point or time series, single or multiple reads
per well), and reads the data
}
\examples{
dat <- tecan_parse(
  system.file(
    "extdata",
    "tecan_time_series_multi_reads.xlsx",
    package = "tecanr"
  )
)

}
