% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tecan_raw.R
\docType{data}
\name{tecan_raw}
\alias{tecan_raw}
\title{Raw data for conversion to long format}
\format{
A list of four tibbles
\describe{
\item{dat_raw_1}{single timepoint and single read per well}
\item{dat_raw_2}{single timepoint and multiple reads per well}
\item{dat_raw_3}{timeseries and single read per well}
\item{dat_raw_4}{timeseries and multiple reads per well}
}
}
\source{
Gleb Ebert
}
\usage{
tecan_raw
}
\description{
Four datasets of Tecan Infinite 200 Pro runs that use different combinations
of two parameters: single timepoint measurement or timeseries of measurements
and single or multiple reads per well
}
\keyword{datasets}
