% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tecan_read_series.R
\name{tecan_read_series}
\alias{tecan_read_series}
\title{Reads a time series of measurements from a tecan plate reader excel file}
\usage{
tecan_read_series(xlsx_file, xlsx_sheet = 1)
}
\arguments{
\item{xlsx_file}{(character) path to Excel file}

\item{xlsx_sheet}{(numeric) number of Excel sheet to read (default: 1)}
}
\value{
(tibble) long format data
}
\description{
Reads a time series of measurements from a tecan plate reader excel file
}
\examples{
tecan_read_series(system.file("extdata", "tecan_series.xlsx", package = "tecanr"))

}
