% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tecan_read_single.R
\name{tecan_read_single}
\alias{tecan_read_single}
\title{Reads a single-timepoint measurement from a tecan plate reader excel file}
\usage{
tecan_read_single(xlsx_file, xlsx_sheet = 1)
}
\arguments{
\item{xlsx_file}{path to Excel file}

\item{xlsx_sheet}{number of Excel sheet to read (default: 1)}
}
\value{
(tibble) long format data
}
\description{
Reads a single-timepoint measurement from a tecan plate reader excel file
}
\examples{
tecan_read_series(system.file("extdata", "tecan_single.xlsx", package = "tecanr"))

}
