% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tecan_read_temperature.R
\name{tecan_read_temperature}
\alias{tecan_read_temperature}
\title{Get time series of temperature values}
\usage{
tecan_read_temperature(xlsx_file, xlsx_sheet = 1)
}
\arguments{
\item{xlsx_file}{(character) path to Excel file}

\item{xlsx_sheet}{(numeric) index of Excel sheet to read (default: 1)}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} containing tidy data
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{\link[=tecan_read_temperature]{tecan_read_temperature()}} was deprecated because it only supported one type
of data (time series with multiple readings per well) and the same result can
be achieved with \code{\link[=tecan_parse]{tecan_parse()}}
}
\examples{
tecan_read_temperature(
  system.file(
    "extdata",
    "tecan_time_series_multiple_reads.xlsx",
    package = "tecanr"
  )
)
# ->
tecan_parse(
  system.file(
    "extdata",
    "tecan_time_series_multiple_reads.xlsx",
    package = "tecanr"
  )
) |>
dplyr::select(time, temp) |>
dplyr::distinct()

}
\keyword{internal}
