% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tecan_single_time_multiple_reads.R
\name{tecan_single_time_multiple_reads}
\alias{tecan_single_time_multiple_reads}
\title{Reads a single time point of multiple OD measurements per well from a tecan
plate reader excel file}
\usage{
tecan_single_time_multiple_reads(dat_raw)
}
\arguments{
\item{dat_raw}{(tibble) a excel sheet}
}
\value{
(tibble) long format data
}
\description{
Reads a single time point of multiple OD measurements per well from a tecan
plate reader excel file
}
