% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tecan_time_series_multiple_reads.R
\name{tecan_time_series_multiple_reads}
\alias{tecan_time_series_multiple_reads}
\title{Reads a time series of multiple OD measurements per well from a tecan plate
reader excel file}
\usage{
tecan_time_series_multiple_reads(dat_raw)
}
\arguments{
\item{dat_raw}{(tibble) a excel sheet}
}
\value{
(tibble) long format data
}
\description{
Reads a time series of multiple OD measurements per well from a tecan plate
reader excel file
}
