% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_multiple_reads.R
\name{time_series_multiple_reads}
\alias{time_series_multiple_reads}
\title{Read a time series of multiple measurements per well}
\usage{
time_series_multiple_reads(dat_raw)
}
\arguments{
\item{dat_raw}{(tibble) an Excel sheet as returned by \code{\link[readxl:read_excel]{readxl::read_xlsx()}}}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} containing tidy data
}
\description{
\code{\link[=time_series_multiple_reads]{time_series_multiple_reads()}} gets a series of Tecan plate reader
measurements from an Excel file when each timepoint contains multiple
readings of each well
}
\examples{
traw$time_series_multiple_reads |> time_series_multiple_reads()

}
