% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_single_reads.R
\name{time_series_single_reads}
\alias{time_series_single_reads}
\title{Read time series of measurements}
\usage{
time_series_single_reads(dat_raw)
}
\arguments{
\item{dat_raw}{(tibble) a excel sheet}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} containing tidy data
}
\description{
\code{\link[=time_series_single_reads]{time_series_single_reads()}} gets a time series of Tecan plate reader
measurements from an Excel file when a single reading was taken of each well
}
\examples{
traw$time_series_single_reads |> time_series_single_reads()

}
