% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tread-package.R
\docType{package}
\name{tread-package}
\alias{tread}
\alias{tread-package}
\title{tread: Read Tecan Infinite 200 Pro Data}
\description{
Read Excel files generated by a Tecan Infinite 200 Pro plate reader. The data is returned in tidy format ready for downstream analysis or plotting.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.gl-eb.me/tread/}
  \item \url{https://github.com/gl-eb/tread}
  \item Report bugs at \url{https://github.com/gl-eb/tread/issues}
}

}
\author{
\strong{Maintainer}: Gleb G. Ebert \email{dev@gl-eb.me} (\href{https://orcid.org/0000-0002-7734-3824}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item ETH Zurich [funder, copyright holder]
}

}
\keyword{internal}
