% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tread_temperature.R
\name{tread_temperature}
\alias{tread_temperature}
\title{Get time series of temperature values}
\usage{
tread_temperature(xlsx_file, xlsx_sheet = 1)
}
\arguments{
\item{xlsx_file}{(character) path to Excel file}

\item{xlsx_sheet}{(numeric) index of Excel sheet to read (default: 1)}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} containing tidy data
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{\link[=tread_temperature]{tread_temperature()}} was deprecated because it only supported one type
of data (time series with multiple readings per well) and the same result can
be achieved with \code{\link[=tparse]{tparse()}}
}
\examples{
tread_temperature(
  system.file(
    "extdata",
    "time_series_multiple_reads.xlsx",
    package = "tread"
  )
)
# ->
tparse(
  system.file(
    "extdata",
    "time_series_multiple_reads.xlsx",
    package = "tread"
  )
) |>
dplyr::select(time, temp) |>
dplyr::distinct()

}
\keyword{internal}
