% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tunite.R
\name{tunite}
\alias{tunite}
\title{Stitch together segments of time series measurements}
\usage{
tunite(xlsx_file, segments = NULL, skip = 0)
}
\arguments{
\item{xlsx_file}{(character) path to Excel file}

\item{segments}{(numeric) number of segments to stitch together, starting
from the last sheet (default: number of sheets)}

\item{skip}{(numeric) number of sheets to skip, starting with the last sheet
(default: 0). Note that empty sheets will be skipped automatically}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} containing tidy data
}
\description{
Parses a Tecan Infinite 200 Pro plate reader Excel file that contains
segments of a time series spread over multiple sheets and automatically
stitches them together accounting for the time gap between them
}
\examples{
tunite(
  system.file(
    "extdata",
    "time_series_segments.xlsx",
    package = "tread"
  )
)

}
