% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tecan_parse.R
\name{tecan_parse}
\alias{tecan_parse}
\title{Parse a Tecan Infinite 200 Pro plate reader file}
\usage{
tecan_parse(xlsx_file, xlsx_sheet = 1)
}
\arguments{
\item{xlsx_file}{(character) path to Excel file}

\item{xlsx_sheet}{(numeric) index of Excel sheet to read from (default: 1)}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} containing tidy data
}
\description{
\code{\link[=tecan_parse]{tecan_parse()}} reads data from a the Excel file produced by a Tecan
Infinite 200 Pro plate reader. It automatically determines the type of
measurement (single time point or time series, single or multiple reads
per well) and will handle the data appropriately.
}
\examples{
dat <- tecan_parse(
  system.file(
    "extdata",
    "tecan_time_series_multiple_reads.xlsx",
    package = "tecanr"
  )
)

}
