% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tecan_read_temperature.R
\name{tecan_read_temperature}
\alias{tecan_read_temperature}
\title{Get time series of temperature values}
\usage{
tecan_read_temperature(xlsx_file, xlsx_sheet = 1)
}
\arguments{
\item{xlsx_file}{(character) path to Excel file}

\item{xlsx_sheet}{(numeric) index of Excel sheet to read (default: 1)}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} containing tidy data
}
\description{
\code{\link[=tecan_read_temperature]{tecan_read_temperature()}} gets the temperature readings corresponding to a
timeseries of Tecan Infinite 200 Pro measurements from an Excel file
}
\examples{
tecan_read_temperature(
  system.file(
    "extdata",
    "tecan_time_series_multi_reads.xlsx",
    package = "tecanr"
  )
)

}
