% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tecan_read_temperature.R
\name{tecan_read_temperature}
\alias{tecan_read_temperature}
\title{Reads the temperature readings corresponding to a timeseries of Tecan
Infinite 200 Pro measurements}
\usage{
tecan_read_temperature(xlsx_file, xlsx_sheet = 1)
}
\arguments{
\item{xlsx_file}{(character) path to Excel file}

\item{xlsx_sheet}{(numeric) number of Excel sheet to read (default: 1)}
}
\description{
Reads the temperature readings corresponding to a timeseries of Tecan
Infinite 200 Pro measurements
}
\examples{
tecan_read_temperature(
  system.file(
    "extdata",
    "tecan_timeSeries_multiReads.xlsx",
    package = "tecanr"
  )
)

}
