% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tecan_time_series_multiple_reads.R
\name{tecan_time_series_multiple_reads}
\alias{tecan_time_series_multiple_reads}
\title{Read a time series of multiple measurements per well}
\usage{
tecan_time_series_multiple_reads(dat_raw)
}
\arguments{
\item{dat_raw}{(tibble) an Excel sheet as returned by \code{\link[readxl:read_excel]{readxl::read_xlsx()}}}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} containing tidy data
}
\description{
\code{\link[=tecan_time_series_multiple_reads]{tecan_time_series_multiple_reads()}} gets a series of Tecan plate reader
measurements from an Excel file when each timepoint contains multiple
readings of each well
}
\examples{
tecan_raw$time_series_multiple_reads |> tecan_time_series_multiple_reads()

}
